+++ docs/conf.py
@@ -21,7 +21,7 @@ sys.path.insert(0, os.path.abspath('..'))
 
 # -- Project information -----------------------------------------------------
 
-project = 'OCDS Kingfisher Views'
+project = 'OCDS Kingfisher Summarize'
 copyright = '2019, Open Contracting Partnership'
 author = 'Open Contracting Partnership'