+++ ocdskingfishersummarize/exceptions.py
@@ -1,6 +1,6 @@
-class KingfisherViewsError(Exception):
+class KingfisherSummarizeError(Exception):
     """Base class for exceptions from within this package"""
 
 
-class AmbiguousSourceError(KingfisherViewsError):
+class AmbiguousSourceError(KingfisherSummarizeError):
     """Raised if a SQL object has an ambiguous source"""