+++ tests/test_add.py
@@ -6,16 +6,16 @@ import pytest
 from click.testing import CliRunner
 from psycopg2 import sql
 
-from ocdskingfisherviews.cli import cli
+from manage import cli
 from tests import assert_bad_argument, assert_log_records, assert_log_running, fixture, noop
 
-command = 'add-view'
+command = 'add'
 
 TABLES = {
     'note',
     'selected_collections',
 
-    # refresh_views
+    # summarize
     'award_documents_summary',
     'award_items_summary',
     'award_suppliers_summary',
@@ -67,8 +67,8 @@ def test_validate_collections(collections, message, caplog):
     assert_log_running(caplog, command)
 
 
-@patch('ocdskingfisherviews.cli.refresh_views', noop)
-@patch('ocdskingfisherviews.cli.field_counts', noop)
+@patch('manage.summary_tables', noop)
+@patch('manage.field_counts', noop)
 @pytest.mark.parametrize('kwargs, name, collections', [
     ({}, 'collection_1', (1,)),
     ({'collections': '1,2'}, 'collection_1_2', (1, 2)),
@@ -92,7 +92,7 @@ def test_command_name(kwargs, name, collections, db, caplog):
         assert_log_records(caplog, command, [
             f'Arguments: collections={collections!r} note=Default name={kwargs.get("name")} tables_only=False',
             f'Added {name}',
-            'Running refresh-views routine',
+            'Running summary-tables routine',
             'Running field-counts routine',
             'Running correct-user-permissions command',
         ])
@@ -237,7 +237,7 @@ def test_command(db, tables_only, tables, views, caplog):
         assert_log_records(caplog, command, [
             f'Arguments: collections=(1,) note=Default name=None tables_only={tables_only!r}',
             'Added collection_1',
-            'Running refresh-views routine',
+            'Running summary-tables routine',
             'Running field-counts routine',
             'Running correct-user-permissions command',
         ])