+++ tests/test_functions.py
@@ -7,7 +7,7 @@ import pytest
 from tests import fixture
 
 
-@patch('ocdskingfisherviews.cli.dependency_graph', dict)
+@patch('manage.dependency_graph', dict)
 @pytest.mark.parametrize('text, expected', [
     ('123.456', decimal.Decimal('123.456')),
     ('123a', None),
@@ -19,7 +19,7 @@ def test_convert_to_numeric(db, text, expected):
         assert value == expected
 
 
-@patch('ocdskingfisherviews.cli.dependency_graph', dict)
+@patch('manage.dependency_graph', dict)
 @pytest.mark.parametrize('text, expected', [
     ('2020-02-29T00:00:00Z', datetime.datetime(2020, 2, 29, 0, 0)),
     ('2020-02-30T00:00:00Z', None),
@@ -32,7 +32,7 @@ def test_convert_to_timestamp(db, text, expected):
         assert value == expected
 
 
-@patch('ocdskingfisherviews.cli.dependency_graph', dict)
+@patch('manage.dependency_graph', dict)
 @pytest.mark.parametrize('scheme, id_, expected', [
     ('prefix', '123', 'prefix-123'),
     ('prefix', None, 'prefix'),